+++ CHANGELOG.md
@@ -10,6 +10,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `block` objective now supports the parameter `ignorecancel` allow counting breaking / placing blocks in protected regions
 - command `variable` to list and modify variables on a variable objective
 ### Changed
+- player variable
+  - `%player%` is now the same as `%player.name%`
+  - `%player.name%` will display the name of the player
+  - `%player.display%` will use the display name used in chat
+  - `%player.uuid%` will display the UUID of the player
 ### Deprecated
 ### Removed
 ### Fixed